requirements.txt
azure-data-tables
azure-kusto-data
polars
concurrent_log_handler
requests